core/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
mysql-connector-python==9.0.0
pillow==10.4.0
pyTelegramBotAPI==4.23.0
qrcode==8.0
requests==2.32.3
urllib3==2.2.3
